mlua

Struct AnyUserData

pub struct AnyUserData<'lua>(/* private fields */);

Handle to an internal Lua userdata for any type that implements UserData.

Similar to std::any::Any, this provides an interface for dynamic type checking via the is and borrow methods.

Internally, instances are stored in a RefCell, to best match the mutable semantics of the Lua language.

Note

This API should only be used when necessary. Implementing UserData already allows defining methods which check the type and acquire a borrow behind the scenes.

Implementations

impl<'lua> AnyUserData<'lua>

pub fn is<T: 'static>(&self) -> bool

Checks whether the type of this userdata is T.

pub fn borrow<T: 'static>(&self) -> Result<Ref<'_, T>>

Borrow this userdata immutably if it is of type T.

Errors

Returns a UserDataBorrowError if the userdata is already mutably borrowed. Returns a UserDataTypeMismatch if the userdata is not of type T.

pub fn borrow_mut<T: 'static>(&self) -> Result<RefMut<'_, T>>

Borrow this userdata mutably if it is of type T.

Errors

Returns a UserDataBorrowMutError if the userdata cannot be mutably borrowed. Returns a UserDataTypeMismatch if the userdata is not of type T.

pub fn take<T: 'static>(&self) -> Result<T>

Takes the value out of this userdata. Sets the special โdestructedโ metatable that prevents any further operations with this userdata.

Keeps associated user values unchanged (they will be collected by Luaโs GC).

pub fn set_user_value<V: IntoLua<'lua>>(&self, v: V) -> Result<()>

Sets an associated value to this AnyUserData.

The value may be any Lua value whatsoever, and can be retrieved with user_value.

This is the same as calling set_nth_user_value with n set to 1.

pub fn user_value<V: FromLua<'lua>>(&self) -> Result<V>

Returns an associated value set by set_user_value.

This is the same as calling nth_user_value with n set to 1.

pub fn set_nth_user_value<V: IntoLua<'lua>>(&self, n: usize, v: V) -> Result<()>

Sets an associated nth value to this AnyUserData.

The value may be any Lua value whatsoever, and can be retrieved with nth_user_value. n starts from 1 and can be up to 65535.

This is supported for all Lua versions. In Lua 5.4 first 7 elements are stored in a most efficient way. For other Lua versions this functionality is provided using a wrapping table.

pub fn nth_user_value<V: FromLua<'lua>>(&self, n: usize) -> Result<V>

Returns an associated nth value set by set_nth_user_value.

n starts from 1 and can be up to 65535.

This is supported for all Lua versions. In Lua 5.4 first 7 elements are stored in a most efficient way. For other Lua versions this functionality is provided using a wrapping table.

pub fn set_named_user_value<V: IntoLua<'lua>>( &self, name: &str, v: V, ) -> Result<()>

Sets an associated value to this AnyUserData by name.

The value can be retrieved with named_user_value.

pub fn named_user_value<V: FromLua<'lua>>(&self, name: &str) -> Result<V>

Returns an associated value by name set by set_named_user_value.

pub fn get_metatable(&self) -> Result<UserDataMetatable<'lua>>

Returns a metatable of this UserData.

Returned UserDataMetatable object wraps the original metatable and provides safe access to its methods.

For T: 'static returned metatable is shared among all instances of type T.

pub fn to_pointer(&self) -> *const c_void

Converts this userdata to a generic C pointer.

There is no way to convert the pointer back to its original value.

Typically this function is used only for hashing and debug information.

impl<'lua> AnyUserData<'lua>

pub fn wrap<T: MaybeSend + 'static>(data: T) -> impl IntoLua<'lua>

Wraps any Rust type, returning an opaque type that implements IntoLua trait.

This function uses Lua::create_any_userdata() under the hood.