mlua

Struct Chunk

pub struct Chunk<'lua, 'a> { /* private fields */ }

Returned from Lua::load and is used to finalize loading and executing Lua main chunks.

Implementations

impl<'lua, 'a> Chunk<'lua, 'a>

pub fn set_name(self, name: impl Into<String>) -> Self

Sets the name of this chunk, which results in more informative error traces.

pub fn set_environment<V: IntoLua<'lua>>(self, env: V) -> Self

Sets the environment of the loaded chunk to the given value.

In Lua >=5.2 main chunks always have exactly one upvalue, and this upvalue is used as the _ENV variable inside the chunk. By default this value is set to the global environment.

Calling this method changes the _ENV upvalue to the value provided, and variables inside the chunk will refer to the given environment rather than the global one.

All global variables (including the standard library!) are looked up in _ENV, so it may be necessary to populate the environment in order for scripts using custom environments to be useful.

pub fn set_mode(self, mode: ChunkMode) -> Self

Sets whether the chunk is text or binary (autodetected by default).

Be aware, Lua does not check the consistency of the code inside binary chunks. Running maliciously crafted bytecode can crash the interpreter.

pub fn set_compiler(self, compiler: Compiler) -> Self

Sets or overwrites a Luau compiler used for this chunk.

See Compiler for details and possible options.

Requires feature = "luau"

pub fn exec(self) -> Result<()>

Execute this chunk of code.

This is equivalent to calling the chunk function with no arguments and no return values.

pub fn eval<R: FromLuaMulti<'lua>>(self) -> Result<R>

Evaluate the chunk as either an expression or block.

If the chunk can be parsed as an expression, this loads and executes the chunk and returns the value that it evaluates to. Otherwise, the chunk is interpreted as a block as normal, and this is equivalent to calling exec.

pub fn call<A: IntoLuaMulti<'lua>, R: FromLuaMulti<'lua>>( self, args: A, ) -> Result<R>

Load the chunk function and call it with the given arguments.

This is equivalent to into_function and calling the resulting function.

pub fn into_function(self) -> Result<Function<'lua>>

Load this chunk into a regular Function.

This simply compiles the chunk without actually executing it.