mlua

Struct Table

pub struct Table<'lua>(/* private fields */);

Handle to an internal Lua table.

Implementations

impl<'lua> Table<'lua>

pub fn set<K: IntoLua<'lua>, V: IntoLua<'lua>>( &self, key: K, value: V, ) -> Result<()>

Sets a key-value pair in the table.

If the value is nil, this will effectively remove the pair.

This might invoke the __newindex metamethod. Use the raw_set method if that is not desired.

Examples

Export a value as a global to make it usable from Lua:

let globals = lua.globals();

globals.set("assertions", cfg!(debug_assertions))?;

lua.load(r#"
    if assertions == true then
        -- ...
    elseif assertions == false then
        -- ...
    else
        error("assertions neither on nor off?")
    end
"#).exec()?;

pub fn get<K: IntoLua<'lua>, V: FromLua<'lua>>(&self, key: K) -> Result<V>

Gets the value associated to key from the table.

If no value is associated to key, returns the nil value.

This might invoke the __index metamethod. Use the raw_get method if that is not desired.

Examples

Query the version of the Lua interpreter:

let globals = lua.globals();

let version: String = globals.get("_VERSION")?;
println!("Lua version: {}", version);

pub fn contains_key<K: IntoLua<'lua>>(&self, key: K) -> Result<bool>

Checks whether the table contains a non-nil value for key.

This might invoke the __index metamethod.

pub fn push<V: IntoLua<'lua>>(&self, value: V) -> Result<()>

Appends a value to the back of the table.

This might invoke the __len and __newindex metamethods.

pub fn pop<V: FromLua<'lua>>(&self) -> Result<V>

Removes the last element from the table and returns it.

This might invoke the __len and __newindex metamethods.

pub fn equals<T: AsRef<Self>>(&self, other: T) -> Result<bool>

Compares two tables for equality.

Tables are compared by reference first. If they are not primitively equals, then mlua will try to invoke the __eq metamethod. mlua will check self first for the metamethod, then other if not found.

Examples

Compare two tables using __eq metamethod:

let table1 = lua.create_table()?;
table1.set(1, "value")?;

let table2 = lua.create_table()?;
table2.set(2, "value")?;

let always_equals_mt = lua.create_table()?;
always_equals_mt.set("__eq", lua.create_function(|_, (_t1, _t2): (Table, Table)| Ok(true))?)?;
table2.set_metatable(Some(always_equals_mt));

assert!(table1.equals(&table1.clone())?);
assert!(table1.equals(&table2)?);

pub fn raw_set<K: IntoLua<'lua>, V: IntoLua<'lua>>( &self, key: K, value: V, ) -> Result<()>

Sets a key-value pair without invoking metamethods.

pub fn raw_get<K: IntoLua<'lua>, V: FromLua<'lua>>(&self, key: K) -> Result<V>

Gets the value associated to key without invoking metamethods.

pub fn raw_insert<V: IntoLua<'lua>>(&self, idx: Integer, value: V) -> Result<()>

Inserts element value at position idx to the table, shifting up the elements from table[idx]. The worst case complexity is O(n), where n is the table length.

pub fn raw_push<V: IntoLua<'lua>>(&self, value: V) -> Result<()>

Appends a value to the back of the table without invoking metamethods.

pub fn raw_pop<V: FromLua<'lua>>(&self) -> Result<V>

Removes the last element from the table and returns it, without invoking metamethods.

pub fn raw_remove<K: IntoLua<'lua>>(&self, key: K) -> Result<()>

Removes a key from the table.

If key is an integer, mlua shifts down the elements from table[key+1], and erases element table[key]. The complexity is O(n) in the worst case, where n is the table length.

For other key types this is equivalent to setting table[key] = nil.

pub fn clear(&self) -> Result<()>

Clears the table, removing all keys and values from array and hash parts, without invoking metamethods.

This method is useful to clear the table while keeping its capacity.

pub fn len(&self) -> Result<Integer>

Returns the result of the Lua # operator.

This might invoke the __len metamethod. Use the raw_len method if that is not desired.

pub fn raw_len(&self) -> usize

Returns the result of the Lua # operator, without invoking the __len metamethod.

pub fn is_empty(&self) -> bool

Returns true if the table is empty, without invoking metamethods.

It checks both the array part and the hash part.

pub fn get_metatable(&self) -> Option<Table<'lua>>

Returns a reference to the metatable of this table, or None if no metatable is set.

Unlike the getmetatable Lua function, this method ignores the __metatable field.

pub fn set_metatable(&self, metatable: Option<Table<'lua>>)

Sets or removes the metatable of this table.

If metatable is None, the metatable is removed (if no metatable is set, this does nothing).

pub fn set_readonly(&self, enabled: bool)

Sets readonly attribute on the table.

Requires feature = "luau"

pub fn is_readonly(&self) -> bool

Returns readonly attribute of the table.

Requires feature = "luau"

pub fn to_pointer(&self) -> *const c_void

Converts this table to a generic C pointer.

Different tables will give different pointers. There is no way to convert the pointer back to its original value.

Typically this function is used only for hashing and debug information.

pub fn pairs<K: FromLua<'lua>, V: FromLua<'lua>>(self) -> TablePairs<'lua, K, V>

Consume this table and return an iterator over the pairs of the table.

This works like the Lua pairs function, but does not invoke the __pairs metamethod.

The pairs are wrapped in a Result, since they are lazily converted to K and V types.

Note

While this method consumes the Table object, it can not prevent code from mutating the table while the iteration is in progress. Refer to the Lua manual for information about the consequences of such mutation.

Examples

Iterate over all globals:

let globals = lua.globals();

for pair in globals.pairs::<Value, Value>() {
    let (key, value) = pair?;
    // ...
}

pub fn for_each<K, V>(&self, f: impl FnMut(K, V) -> Result<()>) -> Result<()>
where K: FromLua<'lua>, V: FromLua<'lua>,

Iterates over the pairs of the table, invoking the given closure on each pair.

This method is similar to Table::pairs, but optimized for performance. It does not invoke the __pairs metamethod.

pub fn sequence_values<V: FromLua<'lua>>(self) -> TableSequence<'lua, V>

Consume this table and return an iterator over all values in the sequence part of the table.

The iterator will yield all values t[1], t[2] and so on, until a nil value is encountered. This mirrors the behavior of Lua’s ipairs function but does not invoke any metamethods.

Note

While this method consumes the Table object, it can not prevent code from mutating the table while the iteration is in progress. Refer to the Lua manual for information about the consequences of such mutation.

Examples

let my_table: Table = lua.load(r#"
    {
        [1] = 4,
        [2] = 5,
        [4] = 7,
        key = 2
    }
"#).eval()?;

let expected = [4, 5];
for (&expected, got) in expected.iter().zip(my_table.sequence_values::<u32>()) {
    assert_eq!(expected, got?);
}