mlua

Trait ToLuaMulti

pub trait ToLuaMulti<'lua> {
    // Required method
    fn to_lua_multi(self, lua: &'lua Lua) -> Result<MultiValue<'lua>>;
}

Trait for types convertible to any number of Lua values.

This is a generalization of ToLua, allowing any number of resulting Lua values instead of just one. Any type that implements ToLua will automatically implement this trait.

Required Methods

fn to_lua_multi(self, lua: &'lua Lua) -> Result<MultiValue<'lua>>

Performs the conversion.