mlua

Enum Value

pub enum Value<'lua> {
    Nil,
    Boolean(bool),
    LightUserData(LightUserData),
    Integer(Integer),
    Number(Number),
    Vector(Vector),
    String(String<'lua>),
    Table(Table<'lua>),
    Function(Function<'lua>),
    Thread(Thread<'lua>),
    UserData(AnyUserData<'lua>),
    Error(Error),
}

A dynamically typed Lua value. The String, Table, Function, Thread, and UserData variants contain handle types into the internal Lua state. It is a logic error to mix handle types between separate Lua instances, and doing so will result in a panic.

Variants

Nil

The Lua value nil.

Boolean(bool)

The Lua value true or false.

LightUserData(LightUserData)

A “light userdata” object, equivalent to a raw pointer.

Integer(Integer)

An integer number.

Any Lua number convertible to a Integer will be represented as this variant.

Number(Number)

A floating point number.

Vector(Vector)

A Luau vector.

String(String<'lua>)

An interned string, managed by Lua.

Unlike Rust strings, Lua strings may not be valid UTF-8.

Table(Table<'lua>)

Reference to a Lua table.

Function(Function<'lua>)

Reference to a Lua function (or closure).

Thread(Thread<'lua>)

Reference to a Lua thread (or coroutine).

UserData(AnyUserData<'lua>)

Reference to a userdata object that holds a custom type which implements UserData. Special builtin userdata types will be represented as other Value variants.

Error(Error)

Error is a special builtin userdata type. When received from Lua it is implicitly cloned.

Implementations

impl<'lua> Value<'lua>

pub const NULL: Value<'static>

A special value (lightuserdata) to represent null value.

It can be used in Lua tables without downsides of nil.

pub const fn type_name(&self) -> &'static str

Returns type name of this value.

pub fn equals<T: AsRef<Self>>(&self, other: T) -> Result<bool>

Compares two values for equality.

Equality comparisons do not convert strings to numbers or vice versa. Tables, Functions, Threads, and Userdata are compared by reference: two objects are considered equal only if they are the same object.

If Tables or Userdata have __eq metamethod then mlua will try to invoke it. The first value is checked first. If that value does not define a metamethod for __eq, then mlua will check the second value. Then mlua calls the metamethod with the two values as arguments, if found.

pub fn to_pointer(&self) -> *const c_void

Converts the value to a generic C pointer.

The value can be a userdata, a table, a thread, a string, or a function; otherwise it returns NULL. Different objects will give different pointers. There is no way to convert the pointer back to its original value.

Typically this function is used only for hashing and debug information.

pub fn to_string(&self) -> Result<StdString>

Converts the value to a string.

If the value has a metatable with a __tostring method, then it will be called to get the result.

pub fn is_nil(&self) -> bool

Returns true if the value is a Nil.

pub fn is_null(&self) -> bool

Returns true if the value is a [NULL].

pub fn is_boolean(&self) -> bool

Returns true if the value is a boolean.

pub fn as_boolean(&self) -> Option<bool>

Cast the value to boolean.

If the value is a Boolean, returns it or None otherwise.

pub fn is_light_userdata(&self) -> bool

Returns true if the value is a LightUserData.

pub fn as_light_userdata(&self) -> Option<LightUserData>

Cast the value to LightUserData.

If the value is a LightUserData, returns it or None otherwise.

pub fn is_integer(&self) -> bool

Returns true if the value is an Integer.

pub fn as_integer(&self) -> Option<Integer>

Cast the value to Integer.

If the value is a Lua Integer, returns it or None otherwise.

pub fn as_i32(&self) -> Option<i32>

Cast the value to i32.

If the value is a Lua Integer, try to convert it to i32 or return None otherwise.

pub fn as_u32(&self) -> Option<u32>

Cast the value to u32.

If the value is a Lua Integer, try to convert it to u32 or return None otherwise.

pub fn as_i64(&self) -> Option<i64>

Cast the value to i64.

If the value is a Lua Integer, try to convert it to i64 or return None otherwise.

pub fn as_u64(&self) -> Option<u64>

Cast the value to u64.

If the value is a Lua Integer, try to convert it to u64 or return None otherwise.

pub fn as_isize(&self) -> Option<isize>

Cast the value to isize.

If the value is a Lua Integer, try to convert it to isize or return None otherwise.

pub fn as_usize(&self) -> Option<usize>

Cast the value to usize.

If the value is a Lua Integer, try to convert it to usize or return None otherwise.

pub fn is_number(&self) -> bool

Returns true if the value is a Lua Number.

pub fn as_number(&self) -> Option<Number>

Cast the value to Number.

If the value is a Lua Number, returns it or None otherwise.

pub fn as_f32(&self) -> Option<f32>

Cast the value to f32.

If the value is a Lua Number, try to convert it to f32 or return None otherwise.

pub fn as_f64(&self) -> Option<f64>

Cast the value to f64.

If the value is a Lua Number, try to convert it to f64 or return None otherwise.

pub fn is_string(&self) -> bool

Returns true if the value is a Lua String.

pub fn as_string(&self) -> Option<&String<'_>>

Cast the value to Lua String.

If the value is a Lua String, returns it or None otherwise.

pub fn as_str(&self) -> Option<&str>

Cast the value to [str].

If the value is a Lua String, try to convert it to [str] or return None otherwise.

pub fn as_string_lossy(&self) -> Option<Cow<'_, str>>

Cast the value to Cow<str>.

If the value is a Lua String, converts it to Cow<str> or returns None otherwise.

pub fn is_table(&self) -> bool

Returns true if the value is a Lua Table.

pub fn as_table(&self) -> Option<&Table<'_>>

Cast the value to Table.

If the value is a Lua Table, returns it or None otherwise.

pub fn is_thread(&self) -> bool

Returns true if the value is a Lua Thread.

pub fn as_thread(&self) -> Option<&Thread<'_>>

Cast the value to Thread.

If the value is a Lua Thread, returns it or None otherwise.

pub fn is_function(&self) -> bool

Returns true if the value is a Lua Function.

pub fn as_function(&self) -> Option<&Function<'_>>

Cast the value to Function.

If the value is a Lua Function, returns it or None otherwise.

pub fn is_userdata(&self) -> bool

Returns true if the value is an AnyUserData.

pub fn as_userdata(&self) -> Option<&AnyUserData<'_>>

Cast the value to AnyUserData.

If the value is an AnyUserData, returns it or None otherwise.